--- src/app/components/admin-dashboard/admin-dashboard.component.scss
@@ -1,41 +1,0 @@
-.dashboard-container {
-    padding: 20px;
-  }
-  
-  h2 {
-    color: #007bff;
-  }
-  
-  .error-message {
-    color: red;
-    font-weight: bold;
-  }
-  
-  .course-table {
-    width: 100%;
-    border-collapse: collapse;
-    margin-top: 20px;
-  }
-  
-  .course-table th, .course-table td {
-    border: 1px solid #ddd;
-    padding: 8px;
-    text-align: left;
-  }
-  
-  .course-table th {
-    background-color: #007bff;
-    color: white;
-  }
-  
-  button {
-    background-color: #28a745;
-    color: white;
-    border: none;
-    padding: 5px 10px;
-    cursor: pointer;
-  }
-  
-  button:hover {
-    background-color: #218838;
-  }
